style: rename to `item-id.ts` and remove reparator variable
